Add keybinding for switching panel views
This is a system defined shortcut for switching
between editor tabs, so let's enable that
functionality for the panel views, whenever
the panel itself is in focus.

We also disable the default "ctrl+w" binding
attached to the "switchWindow" action.
It interferes with the vim window navigation
prefixes. We weren't able to find a good
reassignment for it yet, so we'll leave it
disabled for now.